moved the alert deadline to int64

+++ pkg/apis/data/v1alpha1/dataset_types.go
@@ -128,7 +128,7 @@ type DatasetSpec struct {
 	// ActiveDeadlineSeconds is the deadline of a job for this dataset.
 	// +kubebuilder:default:=600
 	// +kubebuilder:validation:Optional
-	ActiveDeadlineSeconds *int32 `json:"activeDeadlineSeconds,omitempty" protobuf:"varint,15,opt,name=activeDeadlineSeconds"`
+	ActiveDeadlineSeconds *int64 `json:"activeDeadlineSeconds,omitempty" protobuf:"varint,15,opt,name=activeDeadlineSeconds"`
 }
 
 // DatasetStatus defines the observed state of Dataset
