+++ pkg/apis/training/v1alpha1/report_types.go
@@ -188,7 +188,7 @@ type ReportSpec struct {
 	// ActiveDeadlineSeconds is the deadline of a job for this report.
 	// +kubebuilder:default:=600
 	// +kubebuilder:validation:Optional
-	ActiveDeadlineSeconds *int32 `json:"activeDeadlineSeconds,omitempty" protobuf:"varint,10,opt,name=activeDeadlineSeconds"`
+	ActiveDeadlineSeconds *int64 `json:"activeDeadlineSeconds,omitempty" protobuf:"varint,10,opt,name=activeDeadlineSeconds"`
 }
 
 // ReportStatus defines the observed state of the report.
